requirements.txt
flask
openai
python-telegram-bot==20.3